webserver.computations settings
