+++ services/web/server/src/simcore_service_webserver/computation.py
@@ -9,7 +9,7 @@ from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setu
 
 from .computation_comp_tasks_listening_task import setup as setup_comp_tasks_listener
 from .computation_config import CONFIG_SECTION_NAME
-from .computation_config import create_settings as create_computation_settings
+from .computation_settings import create_settings as create_computation_settings
 from .computation_subscribe import setup_rabbitmq_consumer
 
 log = logging.getLogger(__file__)
